refactor: move utils into client crate
