+++ crates/common/src/client/utils.rs
